Changed the structure and implemented CRUD on assist
I changed the structure of the method call in creates methods and update methods. Instead of having it in all the models, I put a method in the BaseModel and added a variable so it can supports all of them and do code reuse. I also deleted the artist model and the artist routes since we don't need those two files